+++ firefly/ui/app_qt.py
@@ -1794,6 +1794,7 @@ class MainWindow(QtWidgets.QMainWindow, VisualiseMixin, CompareMixin, BatchMixin
             ("performance/hyperfly_max_files", self.s_hyperfly_max_files, "spin", int),
             ("performance/hyperfly_max_cores", self.s_hyperfly_max_cores, "spin", int),
             ("performance/hyperfly_max_ram",   self.s_hyperfly_max_ram,   "spin", int),
+            ("performance/hyperfly_load_slots", self.s_hyperfly_load_slots, "spin", int),
             ("performance/hyperfly_gpu_slots", self.s_hyperfly_gpu_slots, "spin", int),
             ("performance/czi_parallel_decode", self.c_czi_parallel,      "check", _bool_cast),
 
@@ -2847,6 +2848,8 @@ class MainWindow(QtWidgets.QMainWindow, VisualiseMixin, CompareMixin, BatchMixin
                 int(self.s_hyperfly_max_cores.value()))
             os.environ["FIREFLY_HYPERFLY_MAX_RAM_GB"] = str(
                 int(self.s_hyperfly_max_ram.value()))
+            os.environ["FIREFLY_HYPERFLY_LOAD_SLOTS"] = str(
+                int(self.s_hyperfly_load_slots.value()))
             os.environ["FIREFLY_HYPERFLY_GPU_SLOTS"] = str(
                 int(self.s_hyperfly_gpu_slots.value()))
             os.environ["FIREFLY_CZI_PARALLEL_DECODE"] = (